Report and XS tags bugfix
The report now consists of a table summary of some of the files generated
StringTie has problems with the output from STAR as only the spliced reads will have the XS tag. Dobin provides a script in the STAR package to add the XS tag to all (and only) the PE reads. This will hopefully fix the problem...